using float64 instead of int64 && add quant_op registry

+++ brocolli/quantization/quantization_layers/base.py
@@ -11,6 +11,6 @@ class BaseOperator:
             threshold_min = 0
 
         x = torch.clamp(x, threshold_min, threshold_max)
-        x = x.to(torch.int64)
+        x = x.to(torch.float64)
 
         return x
